agents/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-core==0.3.15
langchain-community==0.3.5

# Database & Cache
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25

# Vector Database & Embeddings
qdrant-client==1.7.3
sentence-transformers==2.3.1

# Utilities
numpy==1.26.3
python-multipart==0.0.6
aiofiles==23.2.1
